refactor: fold Cloud ngrok helpers into dw ngrok/start
Cloud is a mode of the existing ngrok helper and startDev, not a
parallel ensureHeadlessNgrok + cloudPublicUrls stack. bun dw / run
go back to calling startDev; provision treats the tunnel as the
headless branch of reserved-domain setup.
