+++ scripts/dw/commands/default.ts
@@ -1,9 +1,8 @@
-import { killOwnPorts } from "../helpers/ports.ts";
-import { getCurrentWorktree } from "../helpers/git.ts";
-import { loadRegistry } from "../helpers/registry.ts";
 import { isProvisioned } from "../helpers/entry.ts";
+import { getCurrentWorktree } from "../helpers/git.ts";
+import { killOwnPorts } from "../helpers/ports.ts";
+import { loadRegistry } from "../helpers/registry.ts";
 import { fatal } from "../helpers/shell.ts";
-import { ensureHeadlessNgrok } from "../helpers/ensureHeadlessNgrok.ts";
 import { startDev } from "../helpers/start.ts";
 
 export async function cmdDefault(): Promise<void> {
@@ -15,14 +14,10 @@ export async function cmdDefault(): Promise<void> {
 			fatal("run 'bun dw setup' first to provision this worktree");
 		}
 		killOwnPorts(entry.worktreeNum);
-		startDev(ensureHeadlessNgrok(entry));
+		startDev(entry);
 		return;
 	}
 
 	killOwnPorts(entry?.worktreeNum ?? 1);
-	startDev(
-		ensureHeadlessNgrok(
-			entry ?? { path: cwd, worktreeNum: 1, createdAt: Date.now() },
-		),
-	);
+	startDev(entry ?? { path: cwd, worktreeNum: 1, createdAt: Date.now() });
 }
