+++ scripts/dw/commands/run.ts
@@ -1,10 +1,9 @@
-import { fatal } from "../helpers/shell.ts";
 import { getCurrentWorktree } from "../helpers/git.ts";
 import { isHeadless } from "../helpers/headless.ts";
-import { loadRegistry } from "../helpers/registry.ts";
 import { killOwnPorts } from "../helpers/ports.ts";
+import { loadRegistry } from "../helpers/registry.ts";
+import { fatal } from "../helpers/shell.ts";
 import { startDev } from "../helpers/start.ts";
-import { ensureHeadlessNgrok } from "../helpers/ensureHeadlessNgrok.ts";
 
 export async function cmdRun(): Promise<void> {
 	if (process.env.NODE_ENV === "production") {
@@ -24,5 +23,5 @@ export async function cmdRun(): Promise<void> {
 	}
 
 	killOwnPorts(entry.worktreeNum);
-	startDev(ensureHeadlessNgrok(entry), { allowTmux: false });
+	startDev(entry, { allowTmux: false });
 }
